pyproject.toml
[project]
name = "headhunterdb"
version = "0.1.0"
description = ""
authors = [
    {name = "SergeyRomanov9896",email = "automag2016@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
